+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.1.0 - 2018-12-03
+- **BREAKING** rename `getAddressInfo()` to `getAddressInfoByContact`
+- add new methods: `getAddressInfo()`, `getAddressListInfo()` and `getUserInfo()`
 
-## 0.0.1 - 2018-11-07
-- wip
+## 0.0.4 - 2018-11-15
+- initial